+++ src/aks-preview/azext_aks_preview/managed_cluster_decorator.py
@@ -4,6 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 import copy
+import datetime
 import os
 from types import SimpleNamespace
 from typing import Any, Dict, List, Optional, Tuple, TypeVar, Union
@@ -48,18 +49,19 @@ from knack.prompting import prompt_y_n
 from azext_aks_preview._consts import (
     CONST_AZURE_SERVICE_MESH_MODE_DISABLED,
     CONST_AZURE_SERVICE_MESH_MODE_ISTIO,
-    CONST_EBPF_DATAPLANE_CILIUM,
     CONST_LOAD_BALANCER_SKU_BASIC,
     CONST_MANAGED_CLUSTER_SKU_TIER_FREE,
     CONST_MANAGED_CLUSTER_SKU_TIER_STANDARD,
     CONST_NETWORK_PLUGIN_AZURE,
     CONST_NETWORK_PLUGIN_MODE_OVERLAY,
+    CONST_NETWORK_DATAPLANE_CILIUM,
     CONST_OUTBOUND_TYPE_LOAD_BALANCER,
     CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY,
     CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY,
     CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
     CONST_PRIVATE_DNS_ZONE_NONE,
     CONST_PRIVATE_DNS_ZONE_SYSTEM,
+    CONST_IGNORE_KUBERNETES_DEPRECATIONS,
 )
 from azext_aks_preview._helpers import (
     check_is_private_cluster,
@@ -82,6 +84,8 @@ from azext_aks_preview.agentpool_decorator import (
     AKSPreviewAgentPoolUpdateDecorator,
 )
 from msrestazure.tools import is_valid_resource_id
+
+from dateutil.parser import parse
 
 logger = get_logger(__name__)
 
@@ -467,26 +471,55 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
                     )
         return network_plugin
 
-    def get_network_plugin_mode(self) -> Union[str, None]:
-        """Get the value of network_plugin_mode.
-
-        Note: Currently this parameter does not support being updated.
+    def get_pod_cidr(self) -> Union[str, None]:
+        """Get the value of pod_cidr.
 
         :return: str or None
         """
-        # read the original value passed by the command
-        network_plugin_mode = self.raw_param.get("network_plugin_mode")
         # try to read the property value corresponding to the parameter from the `mc` object
-        if (
-            self.mc and
-            self.mc.network_profile and
-            self.mc.network_profile.network_plugin_mode is not None
-        ):
-            network_plugin_mode = self.mc.network_profile.network_plugin_mode
+        # only read on CREATE as this property can be updated
+        pod_cidr = self.raw_param.get("pod_cidr")
+
+        if self.decorator_mode == DecoratorMode.CREATE:
+            if (
+                self.mc and
+                self.mc.network_profile and
+                self.mc.network_profile.pod_cidr is not None
+            ):
+                pod_cidr = self.mc.network_profile.pod_cidr
+
+        # this parameter does not need dynamic completion
+        # this parameter does not need validation
+        return pod_cidr
+
+    def get_network_plugin_mode(self) -> Union[str, None]:
+        """Get the value of network_plugin_mode.
+
+        :return: str or None
+        """
+        # overwrite if provided by user
+        network_plugin_mode = self.raw_param.get("network_plugin_mode")
+
+        # try to read the property value corresponding to the parameter from the `mc` object
+        # only read on CREATE as this property can be updated
+        if self.decorator_mode == DecoratorMode.CREATE:
+            if (
+                self.mc and
+                self.mc.network_profile and
+                self.mc.network_profile.network_plugin_mode is not None
+            ):
+                network_plugin_mode = self.mc.network_profile.network_plugin_mode
 
         # this parameter does not need dynamic completion
         # this parameter does not need validation
         return network_plugin_mode
+
+    def get_network_dataplane(self) -> Union[str, None]:
+        """Get the value of network_dataplane.
+
+        :return: str or None
+        """
+        return self.raw_param.get("network_dataplane")
 
     def get_enable_cilium_dataplane(self) -> bool:
         """Get the value of enable_cilium_dataplane
@@ -620,6 +653,48 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         # this parameter does not need dynamic completion
         # this parameter does not need validation
         return node_os_upgrade_channel
+
+    def get_upgrade_override_until(self) -> Union[str, None]:
+        """Obtain the value of upgrade_override_until.
+        :return: string or None
+        """
+        # this parameter does not need dynamic completion
+        # this parameter does not need validation
+        return self.raw_param.get("upgrade_override_until")
+
+    def get_upgrade_settings(self) -> Union[List[str], None]:
+        """Obtain the value of upgrade_settings.
+        :return: List[str] or None
+        """
+        # this parameter does not need dynamic completion
+        # this parameter does not need validation
+        upgrade_settings = self.raw_param.get("upgrade_settings")
+        if upgrade_settings is None:
+            return None
+
+        goal_upgrade_settings_list = []
+
+        if upgrade_settings.strip() == "":
+            if self.get_upgrade_override_until():
+                raise MutuallyExclusiveArgumentError(
+                    "Cannot specify --upgrade-override-until when --upgrade-settings is set to empty string. Set only the --upgrade-override-until parameter instead."
+                )
+            return goal_upgrade_settings_list
+
+        input_upgrade_settings_list = [x.strip() for x in upgrade_settings.split(',')]
+
+        supported_upgrade_settings = [
+            CONST_IGNORE_KUBERNETES_DEPRECATIONS,
+        ]
+
+        for s in input_upgrade_settings_list:
+            if s in goal_upgrade_settings_list or s not in supported_upgrade_settings:
+                raise InvalidArgumentValueError(
+                    f"{upgrade_settings} either has duplicates or contains invalid upgrade-settings. Supported settings include, IgnoreKubernetesDeprecations."
+                )
+            goal_upgrade_settings_list.append(s)
+
+        return goal_upgrade_settings_list
 
     def _get_enable_pod_security_policy(self, enable_validation: bool = False) -> bool:
         """Internal function to obtain the value of enable_pod_security_policy.
@@ -2184,7 +2259,16 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         network_profile.network_plugin_mode = self.context.get_network_plugin_mode()
 
         if self.context.get_enable_cilium_dataplane():
-            network_profile.network_dataplane = CONST_EBPF_DATAPLANE_CILIUM
+            # --network-dataplane was introduced with API v20230202preview to replace --enable-cilium-dataplane.
+            # Keep both for backwards compatibility, but validate that the user sets only one of them.
+            if self.context.get_network_dataplane() is not None:
+                raise MutuallyExclusiveArgumentError(
+                    "Cannot specify --enable-cilium-dataplane and "
+                    "--network-dataplane at the same time"
+                )
+            network_profile.network_dataplane = CONST_NETWORK_DATAPLANE_CILIUM
+        else:
+            network_profile.network_dataplane = self.context.get_network_dataplane()
 
         return mc
 
@@ -2577,6 +2661,7 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
             (self.context.get_api_server_authorized_ip_ranges(), None),
             (self.context.get_nodepool_labels(), None),
             (self.context.raw_param.get("enable_workload_identity"), None),
+            (self.context.raw_param.get("upgrade_settings"), None),
         ]
 
     def check_raw_parameters(self):
@@ -2621,6 +2706,22 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
             " or ".join(option_names)
         )
         raise RequiredArgumentMissingError(error_msg)
+
+    def update_network_plugin_settings(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update network plugin settings of network profile for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        network_plugin_mode = self.context.get_network_plugin_mode()
+        if network_plugin_mode:
+            mc.network_profile.network_plugin_mode = network_plugin_mode
+
+        pod_cidr = self.context.get_pod_cidr()
+        if pod_cidr:
+            mc.network_profile.pod_cidr = pod_cidr
+        return mc
 
     def update_outbound_type_in_network_profile(self, mc: ManagedCluster) -> ManagedCluster:
         """Update outbound type of network profile for the ManagedCluster object.
@@ -3099,6 +3200,57 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
             )
         return mc
 
+    def update_upgrade_settings(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update upgrade settings for the ManagedCluster object.
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        existing_until = None
+        if mc.upgrade_settings is not None and mc.upgrade_settings.override_settings is not None and mc.upgrade_settings.override_settings.until is not None:
+            existing_until = mc.upgrade_settings.override_settings.until
+
+        upgrade_settings = self.context.get_upgrade_settings()
+
+        # There is a limitation on differentiating empty list vs. not set in update requests.
+        # In such case, we'll use a workaround here to disable it by setting the until field to the current time, to make the overrides no longer effective.
+        # For now there's only one allowed override so we can return early here.
+        if upgrade_settings is not None and len(upgrade_settings) == 0:
+            if mc.upgrade_settings is not None and mc.upgrade_settings.override_settings is not None and mc.upgrade_settings.override_settings.control_plane_overrides is not None:
+                if mc.upgrade_settings.override_settings.control_plane_overrides == [CONST_IGNORE_KUBERNETES_DEPRECATIONS]:
+                    if existing_until is not None and existing_until.timestamp() > datetime.datetime.utcnow().timestamp():
+                        mc.upgrade_settings.override_settings.until = datetime.datetime.utcnow()
+            return mc
+
+        override_until = self.context.get_upgrade_override_until()
+        upgrade_ignore_kubernetes_deprecations = upgrade_settings is not None and CONST_IGNORE_KUBERNETES_DEPRECATIONS in upgrade_settings
+
+        if upgrade_settings is not None or override_until is not None:
+            if mc.upgrade_settings is None:
+                mc.upgrade_settings = self.models.ClusterUpgradeSettings()
+            if mc.upgrade_settings.override_settings is None:
+                mc.upgrade_settings.override_settings = self.models.UpgradeOverrideSettings()
+            # sets control_plane_overrides
+            if upgrade_ignore_kubernetes_deprecations:
+                if mc.upgrade_settings.override_settings.control_plane_overrides is None:
+                    mc.upgrade_settings.override_settings.control_plane_overrides = []
+                if CONST_IGNORE_KUBERNETES_DEPRECATIONS not in mc.upgrade_settings.override_settings.control_plane_overrides:
+                    mc.upgrade_settings.override_settings.control_plane_overrides.append(CONST_IGNORE_KUBERNETES_DEPRECATIONS)
+            # sets until
+            if override_until is not None:
+                try:
+                    mc.upgrade_settings.override_settings.until = parse(override_until)
+                except Exception:  # pylint: disable=broad-except
+                    raise InvalidArgumentValueError(
+                        f"{override_until} is not a valid datatime format."
+                    )
+            elif upgrade_ignore_kubernetes_deprecations:
+                default_extended_until = datetime.datetime.utcnow() + datetime.timedelta(days=3)
+                if existing_until is None or existing_until.timestamp() < default_extended_until.timestamp():
+                    mc.upgrade_settings.override_settings.until = default_extended_until
+
+        return mc
+
     def update_mc_profile_preview(self) -> ManagedCluster:
         """The overall controller used to update the preview ManagedCluster profile.
 
@@ -3130,6 +3282,8 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_creation_data(mc)
         # update linux profile
         mc = self.update_linux_profile(mc)
+        # update network profile
+        mc = self.update_network_plugin_settings(mc)
         # update outbound type
         mc = self.update_outbound_type_in_network_profile(mc)
         # update kube proxy config
@@ -3140,5 +3294,7 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_node_resource_group_profile(mc)
         # update auto upgrade profile
         mc = self.update_auto_upgrade_profile(mc)
+        # update auto upgrade profile
+        mc = self.update_upgrade_settings(mc)
 
         return mc