+++ src/containerapp/azext_containerapp/_clients.py
@@ -16,7 +16,7 @@ from knack.log import get_logger
 logger = get_logger(__name__)
 
 PREVIEW_API_VERSION = "2022-11-01-preview"
-CURRENT_API_VERSION = "2022-10-01"
+CURRENT_API_VERSION = "2022-11-01-preview"
 JOBS_API_VERSION = "2022-11-01-preview"
 POLLING_TIMEOUT = 600  # how many seconds before exiting
 POLLING_SECONDS = 2  # how many seconds between requests
@@ -858,6 +858,39 @@ class ManagedEnvironmentClient():
 
         r = send_raw_request(cmd.cli_ctx, "POST", request_url)
         return r.json()
+
+
+class WorkloadProfileClient():
+    @classmethod
+    def list_supported(cls, cmd, location):
+        management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
+        api_version = CURRENT_API_VERSION
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        url_fmt = "{}/subscriptions/{}/providers/Microsoft.App/locations/{}/availableManagedEnvironmentsWorkloadProfileTypes?api-version={}"
+        request_url = url_fmt.format(
+            management_hostname.strip('/'),
+            sub_id,
+            location,
+            api_version)
+
+        r = send_raw_request(cmd.cli_ctx, "GET", request_url)
+        return r.json().get("value")
+
+    @classmethod
+    def list(cls, cmd, resource_group_name, env_name):
+        management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
+        api_version = CURRENT_API_VERSION
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        url_fmt = "{}/subscriptions/{}/resourcegroups/{}/providers/Microsoft.App/managedEnvironments/{}/workloadProfileStates?api-version={}"
+        request_url = url_fmt.format(
+            management_hostname.strip('/'),
+            sub_id,
+            resource_group_name,
+            env_name,
+            api_version)
+
+        r = send_raw_request(cmd.cli_ctx, "GET", request_url)
+        return r.json().get("value")
 
 
 class ContainerAppsJobClient():