+++ src/containerapp/azext_containerapp/commands.py
@@ -132,6 +132,10 @@ def load_command_table(self, _):
         g.custom_command('set', 'set_ingress_traffic', exception_handler=ex_handler_factory())
         g.custom_show_command('show', 'show_ingress_traffic')
 
+    with self.command_group('containerapp ingress sticky-sessions') as g:
+        g.custom_command('set', 'set_ingress_sticky_session', exception_handler=ex_handler_factory())
+        g.custom_show_command('show', 'show_ingress_sticky_session')
+
     with self.command_group('containerapp ingress access-restriction') as g:
         g.custom_command('set', 'set_ip_restriction', exception_handler=ex_handler_factory())
         g.custom_command('remove', 'remove_ip_restriction')
@@ -198,3 +202,10 @@ def load_command_table(self, _):
 
     with self.command_group('containerapp compose') as g:
         g.custom_command('create', 'create_containerapps_from_compose')
+
+    with self.command_group('containerapp env workload-profile', is_preview=True) as g:
+        g.custom_command('list-supported', 'list_supported_workload_profiles')
+        g.custom_command('list', 'list_workload_profiles')
+        g.custom_show_command('show', 'show_workload_profile')
+        g.custom_command('set', 'set_workload_profile')
+        g.custom_command('delete', 'delete_workload_profile')