+++ src/spring/azext_spring/custom.py
@@ -21,7 +21,7 @@ from ._stream_utils import stream_logs
 from azure.mgmt.core.tools import (parse_resource_id, is_valid_resource_id)
 from ._utils import (get_portal_uri, get_spring_sku, get_proxy_api_endpoint, BearerAuth)
 from knack.util import CLIError
-from .vendored_sdks.appplatform.v2023_01_01_preview import models, AppPlatformManagementClient
+from .vendored_sdks.appplatform.v2023_03_01_preview import models, AppPlatformManagementClient
 from knack.log import get_logger
 from azure.cli.core.azclierror import ClientRequestError, FileOperationError, InvalidArgumentValueError, ResourceNotFoundError
 from azure.cli.core.commands.client_factory import get_mgmt_service_client, get_subscription_id
@@ -243,7 +243,8 @@ def app_append_persistent_storage(cmd, client, resource_group, service, name,
                                   mount_path,
                                   share_name=None,
                                   mount_options=None,
-                                  read_only=None):
+                                  read_only=None,
+                                  enable_sub_path=None):
     resource = client.services.get(resource_group, service)
     storage_id = None
     if resource.sku.tier.upper() == 'STANDARDGEN2':
@@ -264,7 +265,8 @@ def app_append_persistent_storage(cmd, client, resource_group, service, name,
         share_name=share_name,
         mount_path=mount_path,
         mount_options=mount_options,
-        read_only=read_only)
+        read_only=read_only,
+        enable_sub_path=enable_sub_path)
 
     custom_persistent_disks.append(
         models.CustomPersistentDiskResource(
@@ -534,9 +536,9 @@ def app_tail_log_internal(cmd, client, resource_group, service, name,
 
     if timeout:
         t.join(timeout=timeout)
-
-    while t.is_alive():
-        sleep(5)  # so that ctrl+c can stop the command
+    else:
+        while t.is_alive():
+            sleep(5)  # so that ctrl+c can stop the command
 
     if exceptions:
         raise exceptions[0]